+++ pyatv/support/device_info.py
