+++ tests/protocols/mrp/test_mrp.py
@@ -6,9 +6,9 @@ import pytest
 
 from pyatv.const import OperatingSystem
 from pyatv.core import mdns
-from pyatv.core.device_info import lookup_version
 from pyatv.interface import DeviceInfo
 from pyatv.protocols.mrp import device_info, scan
+from pyatv.support.device_info import lookup_version
 
 MRP_SERVICE = "_mediaremotetv._tcp.local"
